fix(event-picker): fix onClick trigger multiple error
